refactor(image-controller): split watcher unit tests into a sibling file
`watcher.rs` crossed the repository's 51200-byte Rust source guard
(51951 bytes). Move the test module into `watcher_tests.rs` behind the
`#[path]` convention `run_dirs_tests.rs` and `graph_warmer_tests.rs`
already use — it stays a child module of `watcher`, so the private
sentinel parsers remain reachable and no visibility was widened.
